task_4/recepie_book/requirements.txt
asgiref==3.5.0
Django==3.2.8
django-js-asset==2.0.0
django-mptt==0.13.4
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
